reorganised application project hierarchy
