+++ Marketplace.Application/Ad/Contracts/V1/AddPicture.cs
@@ -1,4 +1,6 @@
-﻿namespace Marketplace.Application.Contracts.V1;
+﻿using Marketplace.Application.Shared;
+
+namespace Marketplace.Application.Ad.Contracts.V1;
 public class AddPicture : AbstractContract
 {
     public int Height { get; set; }
