+++ Marketplace.Application/Ad/Contracts/V1/Create.cs
@@ -1,0 +1,8 @@
+﻿using Marketplace.Application.Shared;
+
+namespace Marketplace.Application.Ad.Contracts.V1;
+
+public class Create : AbstractContract
+{
+    public Guid OwnerId { get; set; }
+}
