+++ Marketplace.Application/Ad/Contracts/V1/RequestToPublish.cs
@@ -1,0 +1,4 @@
+﻿using Marketplace.Application.Shared;
+
+namespace Marketplace.Application.Ad.Contracts.V1;
+public class RequestToPublish : AbstractContract { }
