+++ Marketplace.Application/Ad/Contracts/V1/SetTitle.cs
@@ -1,4 +1,6 @@
-﻿namespace Marketplace.Application.Contracts.V1;
+﻿using Marketplace.Application.Shared;
+
+namespace Marketplace.Application.Ad.Contracts.V1;
 public class SetTitle : AbstractContract
 {
     public string Title { get; set; } = string.Empty;
