--- Marketplace.Application/Contracts/V1/Create.cs
@@ -1,6 +1,0 @@
-﻿namespace Marketplace.Application.Contracts.V1;
-
-public class Create : AbstractContract
-{
-    public Guid OwnerId {  get; set; }
-}
