--- Marketplace.Application/Contracts/V1/RequestToPublish.cs
@@ -1,2 +1,0 @@
-﻿namespace Marketplace.Application.Contracts.V1;
-public class RequestToPublish : AbstractContract {}
