--- Marketplace.Application/Services/IApplicationService.cs
@@ -1,7 +1,0 @@
-﻿using Marketplace.Application.Contracts.V1;
-
-namespace Marketplace.Application.Services.Ad;
-public interface IApplicationService
-{
-    Task Handle(AbstractContract command);
-}
