+++ Marketplace.Application/Shared/AbstractContract.cs
@@ -1,4 +1,4 @@
-﻿namespace Marketplace.Application.Contracts.V1;
+﻿namespace Marketplace.Application.Shared;
 public abstract class AbstractContract
 {
     public Guid Id { get; set; }
