+++ Marketplace.Application/Shared/Services/IApplicationService.cs
@@ -1,0 +1,5 @@
+﻿namespace Marketplace.Application.Shared.Services;
+public interface IApplicationService
+{
+    Task Handle(AbstractContract command);
+}
